+++ src/Models/User.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Althinect\FilamentSpatieRolesPermissions\Models;
+
+use App\Models\User as ModelsUser;
+
+class User extends ModelsUser
+{
+
+}